ocr-service/requirements.txt
opencv-python
fastapi[standard]
pytesseract
Pillow